Change: Establish `storage-v2` as the default and remove `RaftStorage`
Prior to this commit, the new storage API—comprising `RaftLogStorage`
and `RaftStateMachine`—was only accessible when the `storage-v2` feature
flag was explicitly enabled.

Since this commit, the new storage API is now the standard, rendering
the previous `RaftStorage` API obsolete and consequently, it has been
removed from the codebase.

Additional Updates:

- The example storage implementations `memstore`, `rocksstore`, and
  `sledstore` have been reorganized into the `stores/` subdirectory for
  better structure and clarity.

- `RaftStorage` are removed from docs comments and replaced with the new
  two traits.
