+++ .github/workflows/ci.yaml
@@ -143,9 +143,9 @@ jobs:
       fail-fast: false
       matrix:
         include:
-          - store: "memstore"
-          - store: "rocksstore"
-          - store: "sledstore"
+          - store: "stores/memstore"
+          - store: "stores/rocksstore"
+          - store: "stores/sledstore"
 
     steps:
       - name: Setup | Checkout
@@ -173,40 +173,6 @@ jobs:
           RUST_BACKTRACE: full
           OPENRAFT_STORE_DEFENSIVE: on
 
-
-  # Test external stores that enable different openraft features.
-  external-stores:
-    runs-on: ubuntu-latest
-
-    strategy:
-      fail-fast: false
-      matrix:
-        include:
-          - store: "stores/rocksstore-v2"
-
-    steps:
-      - name: Setup | Checkout
-        uses: actions/checkout@v2
-
-
-      - name: Setup | Toolchain
-        uses: actions-rs/toolchain@v1.0.6
-        with:
-          toolchain: "nightly"
-          override: true
-
-
-      - name: Unit Tests
-        uses: actions-rs/cargo@v1
-        with:
-          command: test
-          args: --manifest-path "${{ matrix.store }}/Cargo.toml"
-        env:
-          # Parallel tests block each other and result in timeout.
-          RUST_TEST_THREADS: 2
-          RUST_LOG: debug
-          RUST_BACKTRACE: full
-          OPENRAFT_STORE_DEFENSIVE: on
 
   # Test external crate.
   # cluster_benchmark is a standalone crate for benchmarking openraft cluster.
