+++ cluster_benchmark/Cargo.toml
@@ -19,7 +19,7 @@ repository = "https://github.com/datafuselabs/openraft"
 [dependencies]
 
 [dev-dependencies]
-openraft           = { path="../openraft", version = "0.9.0", features = ["serde", "storage-v2"] }
+openraft           = { path="../openraft", version = "0.10.0", features = ["serde"] }
 
 anyhow = "1.0.63"
 maplit = "1.0.2"
