+++ examples/memstore/Cargo.toml
@@ -15,7 +15,7 @@ license = "MIT OR Apache-2.0"
 repository = "https://github.com/datafuselabs/openraft"
 
 [dependencies]
-openraft = { path = "../../openraft", features = ["serde", "storage-v2"] }
+openraft = { path = "../../openraft", features = ["serde"] }
 
 tokio = { version = "1.0", default-features = false, features = ["sync"] }
