+++ examples/raft-kv-memstore-generic-snapshot-data/Cargo.toml
@@ -17,7 +17,7 @@ repository = "https://github.com/datafuselabs/openraft"
 
 [dependencies]
 memstore = { path = "../memstore", features = [] }
-openraft = { path = "../../openraft", features = ["serde", "storage-v2", "generic-snapshot-data"] }
+openraft = { path = "../../openraft", features = ["serde", "generic-snapshot-data"] }
 
 clap = { version = "4.1.11", features = ["derive", "env"] }
 reqwest = { version = "0.11.9", features = ["json"] }
