+++ examples/raft-kv-memstore-singlethreaded/Cargo.toml
@@ -16,7 +16,7 @@ license = "MIT OR Apache-2.0"
 repository = "https://github.com/datafuselabs/openraft"
 
 [dependencies]
-openraft = { path = "../../openraft", features = ["serde", "storage-v2", "singlethreaded"] }
+openraft = { path = "../../openraft", features = ["serde", "singlethreaded"] }
 
 clap = { version = "4.1.11", features = ["derive", "env"] }
 reqwest = { version = "0.11.9", features = ["json"] }
