--- rocksstore/README.md
@@ -1,5 +1,0 @@
-# openraft-rocksstore
-
-This is an example `RaftStorage` implementation with [`rocksdb`](https://docs.rs/rocksdb/latest/rocksdb/) based on [openraft-0.8](https://github.com/datafuselabs/openraft/tree/release-0.8).
-
-This crate is built mainly for testing or demonstrating purpose.:)
