--- sledstore/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "openraft-sledstore"
-description = "A sled based implementation of the `openraft::RaftStorage` trait."
-documentation = "https://docs.rs/openraft-sledstore"
-readme = "README.md"
-
-version       = { workspace = true }
-edition       = { workspace = true }
-authors       = { workspace = true }
-categories    = { workspace = true }
-homepage      = { workspace = true }
-keywords      = { workspace = true }
-license       = { workspace = true }
-repository    = { workspace = true }
-
-[dependencies]
-openraft = { path= "../openraft", version = "0.9.0", features=["serde"] }
-
-sled = "0.34.7"
-byteorder = "1.4.3"
-serde = { version = "1.0.114", features = ["derive"] }
-serde_json = "1.0.57"
-async-std = { version = "1.12.0", features = ["attributes", "tokio1"] }
-tracing = "0.1.29"
-
-[dev-dependencies]
-tempfile = { version = "3.4.0" }
-
-[package.metadata.docs.rs]
-all-features = true
