--- sledstore/README.md
@@ -1,5 +1,0 @@
-# openraft-sledstore
-
-This is an example `RaftStorage` implementation with [`sled`](https://github.com/spacejam/sled) based on [openraft-0.8](https://github.com/datafuselabs/openraft/tree/release-0.8).
-
-This crate is built mainly for testing or demonstrating purpose.:)
