+++ stores/README.md
@@ -1,4 +1,6 @@
-Storage implementations.
+Example Storage implementations.
 
-These crates are not members of the workspace, because they enable different features.
-E.g., `rocksstore-v2` enables `storage-v2` feature.
+- `memstore` is in-memory storage and is used by the test cases `./tests`.
+
+If a crate has different feature flags enabled, it must not be members of the workspace.
+A feature flag will be enabled for the entire workspace if a member crate enables it.
