+++ stores/memstore/README.md
