+++ stores/memstore/src/lib.rs
@@ -11,16 +11,18 @@ use std::ops::RangeBounds;
 use std::sync::Arc;
 use std::sync::Mutex;
 
+use openraft::storage::LogFlushed;
 use openraft::storage::LogState;
 use openraft::storage::RaftLogReader;
+use openraft::storage::RaftLogStorage;
 use openraft::storage::RaftSnapshotBuilder;
+use openraft::storage::RaftStateMachine;
 use openraft::storage::Snapshot;
 use openraft::Entry;
 use openraft::EntryPayload;
 use openraft::LogId;
 use openraft::OptionalSend;
 use openraft::RaftLogId;
-use openraft::RaftStorage;
 use openraft::RaftTypeConfig;
 use openraft::SnapshotMeta;
 use openraft::StorageError;
@@ -117,8 +119,31 @@ pub enum BlockOperation {
     PurgeLog,
 }
 
-/// An in-memory storage system implementing the `RaftStorage` trait.
-pub struct MemStore {
+/// Block operations for testing purposes.
+#[derive(Clone, Debug, Default)]
+pub struct BlockConfig {
+    inner: Arc<Mutex<BTreeMap<BlockOperation, Duration>>>,
+}
+
+impl BlockConfig {
+    /// Block an operation for testing purposes.
+    pub fn set_blocking(&self, block: BlockOperation, d: Duration) {
+        self.inner.lock().unwrap().insert(block, d);
+    }
+
+    /// Get the blocking flag for an operation.
+    pub fn get_blocking(&self, block: &BlockOperation) -> Option<Duration> {
+        self.inner.lock().unwrap().get(block).cloned()
+    }
+
+    /// Clear a blocking flag for an operation.
+    pub fn clear_blocking(&mut self, block: BlockOperation) {
+        self.inner.lock().unwrap().remove(&block);
+    }
+}
+
+/// An in-memory log storage implementing the `RaftLogStorage` trait.
+pub struct MemLogStore {
     last_purged_log_id: RwLock<Option<LogId<MemNodeId>>>,
 
     committed: RwLock<Option<LogId<MemNodeId>>>,
@@ -126,42 +151,52 @@ pub struct MemStore {
     /// The Raft log. Logs are stored in serialized json.
     log: RwLock<BTreeMap<u64, String>>,
 
-    /// The Raft state machine.
-    sm: RwLock<MemStoreStateMachine>,
-
     /// Block operations for testing purposes.
-    block: Mutex<BTreeMap<BlockOperation, Duration>>,
+    block: BlockConfig,
 
     /// The current hard state.
     vote: RwLock<Option<Vote<MemNodeId>>>,
-
-    snapshot_idx: Arc<Mutex<u64>>,
-
-    /// The current snapshot.
-    current_snapshot: RwLock<Option<MemStoreSnapshot>>,
 }
 
-impl MemStore {
-    /// Create a new `MemStore` instance.
-    pub fn new() -> Self {
+impl MemLogStore {
+    pub fn new(block: BlockConfig) -> Self {
         let log = RwLock::new(BTreeMap::new());
-        let sm = RwLock::new(MemStoreStateMachine::default());
-        let current_snapshot = RwLock::new(None);
 
         Self {
             last_purged_log_id: RwLock::new(None),
             committed: RwLock::new(None),
             log,
-            sm,
-            block: Mutex::new(BTreeMap::new()),
+            block,
             vote: RwLock::new(None),
-            snapshot_idx: Arc::new(Mutex::new(0)),
-            current_snapshot,
         }
     }
+}
 
-    pub async fn new_async() -> Arc<Self> {
-        Arc::new(Self::new())
+/// An in-memory key-value storage implementing the `RaftStateMachine` trait.
+pub struct MemStateMachine {
+    /// The Raft state machine.
+    sm: RwLock<MemStoreStateMachine>,
+
+    snapshot_idx: Arc<Mutex<u64>>,
+
+    /// The current snapshot.
+    current_snapshot: RwLock<Option<MemStoreSnapshot>>,
+
+    /// Block operations for testing purposes.
+    pub block: BlockConfig,
+}
+
+impl MemStateMachine {
+    pub fn new(block: BlockConfig) -> Self {
+        let sm = RwLock::new(MemStoreStateMachine::default());
+        let current_snapshot = RwLock::new(None);
+
+        Self {
+            sm,
+            snapshot_idx: Arc::new(Mutex::new(0)),
+            current_snapshot,
+            block,
+        }
     }
 
     /// Remove the current snapshot.
@@ -182,30 +217,17 @@ impl MemStore {
         let mut sm = self.sm.write().await;
         *sm = MemStoreStateMachine::default();
     }
-
-    /// Block an operation for testing purposes.
-    pub fn set_blocking(&self, block: BlockOperation, d: Duration) {
-        self.block.lock().unwrap().insert(block, d);
-    }
-
-    /// Get the blocking flag for an operation.
-    pub fn get_blocking(&self, block: &BlockOperation) -> Option<Duration> {
-        self.block.lock().unwrap().get(block).cloned()
-    }
-
-    /// Clear a blocking flag for an operation.
-    pub fn clear_blocking(&mut self, block: BlockOperation) {
-        self.block.lock().unwrap().remove(&block);
-    }
 }
 
-impl Default for MemStore {
-    fn default() -> Self {
-        Self::new()
-    }
+pub fn new_mem_store() -> (Arc<MemLogStore>, Arc<MemStateMachine>) {
+    let block = BlockConfig::default();
+    (
+        Arc::new(MemLogStore::new(block.clone())),
+        Arc::new(MemStateMachine::new(block)),
+    )
 }
 
-impl RaftLogReader<TypeConfig> for Arc<MemStore> {
+impl RaftLogReader<TypeConfig> for Arc<MemLogStore> {
     async fn try_get_log_entries<RB: RangeBounds<u64> + Clone + Debug + OptionalSend>(
         &mut self,
         range: RB,
@@ -223,14 +245,14 @@ impl RaftLogReader<TypeConfig> for Arc<MemStore> {
     }
 }
 
-impl RaftSnapshotBuilder<TypeConfig> for Arc<MemStore> {
+impl RaftSnapshotBuilder<TypeConfig> for Arc<MemStateMachine> {
     #[tracing::instrument(level = "trace", skip(self))]
     async fn build_snapshot(&mut self) -> Result<Snapshot<TypeConfig>, StorageError<MemNodeId>> {
         let data;
         let last_applied_log;
         let last_membership;
 
-        if let Some(d) = self.get_blocking(&BlockOperation::DelayBuildingSnapshot) {
+        if let Some(d) = self.block.get_blocking(&BlockOperation::DelayBuildingSnapshot) {
             tracing::info!(?d, "delay snapshot build");
             tokio::time::sleep(d).await;
         }
@@ -243,7 +265,7 @@ impl RaftSnapshotBuilder<TypeConfig> for Arc<MemStore> {
             last_applied_log = sm.last_applied_log;
             last_membership = sm.last_membership.clone();
 
-            if let Some(d) = self.get_blocking(&BlockOperation::BuildSnapshot) {
+            if let Some(d) = self.block.get_blocking(&BlockOperation::BuildSnapshot) {
                 tracing::info!(?d, "blocking snapshot build");
                 tokio::time::sleep(d).await;
             }
@@ -288,7 +310,9 @@ impl RaftSnapshotBuilder<TypeConfig> for Arc<MemStore> {
     }
 }
 
-impl RaftStorage<TypeConfig> for Arc<MemStore> {
+impl RaftLogStorage<TypeConfig> for Arc<MemLogStore> {
+    type LogReader = Self;
+
     async fn get_log_state(&mut self) -> Result<LogState<TypeConfig>, StorageError<MemNodeId>> {
         let log = self.log.read().await;
         let last_serialized = log.iter().next_back().map(|(_, ent)| ent);
@@ -315,6 +339,10 @@ impl RaftStorage<TypeConfig> for Arc<MemStore> {
         })
     }
 
+    async fn get_log_reader(&mut self) -> Self::LogReader {
+        self.clone()
+    }
+
     #[tracing::instrument(level = "trace", skip(self))]
     async fn save_vote(&mut self, vote: &Vote<MemNodeId>) -> Result<(), StorageError<MemNodeId>> {
         tracing::debug!(?vote, "save_vote");
@@ -339,15 +367,22 @@ impl RaftStorage<TypeConfig> for Arc<MemStore> {
         Ok(*self.committed.read().await)
     }
 
-    async fn last_applied_state(
-        &mut self,
-    ) -> Result<(Option<LogId<MemNodeId>>, StoredMembership<MemNodeId, ()>), StorageError<MemNodeId>> {
-        let sm = self.sm.read().await;
-        Ok((sm.last_applied_log, sm.last_membership.clone()))
+    #[tracing::instrument(level = "trace", skip_all)]
+    async fn append<I>(&mut self, entries: I, callback: LogFlushed<TypeConfig>) -> Result<(), StorageError<MemNodeId>>
+    where I: IntoIterator<Item = Entry<TypeConfig>> + OptionalSend {
+        let mut log = self.log.write().await;
+        for entry in entries {
+            let s =
+                serde_json::to_string(&entry).map_err(|e| StorageIOError::write_log_entry(*entry.get_log_id(), &e))?;
+            log.insert(entry.log_id.index, s);
+        }
+
+        callback.log_io_completed(Ok(()));
+        Ok(())
     }
 
     #[tracing::instrument(level = "debug", skip(self))]
-    async fn delete_conflict_logs_since(&mut self, log_id: LogId<MemNodeId>) -> Result<(), StorageError<MemNodeId>> {
+    async fn truncate(&mut self, log_id: LogId<MemNodeId>) -> Result<(), StorageError<MemNodeId>> {
         tracing::debug!("delete_log: [{:?}, +oo)", log_id);
 
         {
@@ -363,10 +398,10 @@ impl RaftStorage<TypeConfig> for Arc<MemStore> {
     }
 
     #[tracing::instrument(level = "debug", skip_all)]
-    async fn purge_logs_upto(&mut self, log_id: LogId<MemNodeId>) -> Result<(), StorageError<MemNodeId>> {
+    async fn purge(&mut self, log_id: LogId<MemNodeId>) -> Result<(), StorageError<MemNodeId>> {
         tracing::debug!("purge_log_upto: {:?}", log_id);
 
-        if let Some(d) = self.get_blocking(&BlockOperation::PurgeLog) {
+        if let Some(d) = self.block.get_blocking(&BlockOperation::PurgeLog) {
             tracing::info!(?d, "block purging log");
             tokio::time::sleep(d).await;
         }
@@ -388,25 +423,25 @@ impl RaftStorage<TypeConfig> for Arc<MemStore> {
 
         Ok(())
     }
+}
 
-    #[tracing::instrument(level = "trace", skip(self, entries))]
-    async fn append_to_log<I>(&mut self, entries: I) -> Result<(), StorageError<MemNodeId>>
-    where I: IntoIterator<Item = Entry<TypeConfig>> + OptionalSend {
-        let mut log = self.log.write().await;
-        for entry in entries {
-            let s =
-                serde_json::to_string(&entry).map_err(|e| StorageIOError::write_log_entry(*entry.get_log_id(), &e))?;
-            log.insert(entry.log_id.index, s);
-        }
-        Ok(())
+impl RaftStateMachine<TypeConfig> for Arc<MemStateMachine> {
+    type SnapshotBuilder = Self;
+
+    async fn applied_state(
+        &mut self,
+    ) -> Result<(Option<LogId<MemNodeId>>, StoredMembership<MemNodeId, ()>), StorageError<MemNodeId>> {
+        let sm = self.sm.read().await;
+        Ok((sm.last_applied_log, sm.last_membership.clone()))
     }
 
     #[tracing::instrument(level = "trace", skip(self, entries))]
-    async fn apply_to_state_machine(
-        &mut self,
-        entries: &[Entry<TypeConfig>],
-    ) -> Result<Vec<ClientResponse>, StorageError<MemNodeId>> {
-        let mut res = Vec::with_capacity(entries.len());
+    async fn apply<I>(&mut self, entries: I) -> Result<Vec<ClientResponse>, StorageError<MemNodeId>>
+    where
+        I: IntoIterator<Item = Entry<TypeConfig>> + OptionalSend,
+        I::IntoIter: OptionalSend,
+    {
+        let mut res = Vec::new();
 
         let mut sm = self.sm.write().await;
 
@@ -435,6 +470,10 @@ impl RaftStorage<TypeConfig> for Arc<MemStore> {
             };
         }
         Ok(res)
+    }
+
+    async fn get_snapshot_builder(&mut self) -> Self::SnapshotBuilder {
+        self.clone()
     }
 
     #[tracing::instrument(level = "trace", skip(self))]
@@ -493,16 +532,5 @@ impl RaftStorage<TypeConfig> for Arc<MemStore> {
             }
             None => Ok(None),
         }
-    }
-
-    type LogReader = Self;
-    type SnapshotBuilder = Self;
-
-    async fn get_log_reader(&mut self) -> Self::LogReader {
-        self.clone()
-    }
-
-    async fn get_snapshot_builder(&mut self) -> Self::SnapshotBuilder {
-        self.clone()
     }
 }
