+++ stores/rocksstore/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "openraft-rocksstore-v2"
-description = "A rocksdb based implementation of the `openraft::RaftStorage` trait."
+name = "openraft-rocksstore"
+description = "A rocksdb based implementation of the `openraft::RaftLogStorage` and `openraft::RaftStateMachine` trait."
 documentation = "https://docs.rs/openraft-rocksstore"
 readme = "README.md"
 
@@ -18,7 +18,7 @@ repository = "https://github.com/datafuselabs/openraft"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-openraft = { path= "../../openraft", version = "0.9.0", features=["serde", "storage-v2"] }
+openraft = { path= "../../openraft", version = "0.10.0", features=["serde"] }
 
 rocksdb = "0.20.1"
 rand = "*"
