+++ stores/rocksstore/README.md
@@ -1,4 +1,4 @@
-# openraft-rocksstore-v2
+# openraft-rocksstore
 
 This is an example of v2 storage [`RaftLogStorage`] and [`RaftStateMachine`] implementation
 with [`rocksdb`](https://docs.rs/rocksdb/latest/rocksdb/) based on [openraft-0.8](https://github.com/datafuselabs/openraft/tree/release-0.8).
