+++ stores/sledstore/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "openraft-rocksstore"
-description = "A rocksdb based implementation of the `openraft::RaftStorage` trait."
-documentation = "https://docs.rs/openraft-rocksstore"
+name = "openraft-sledstore"
+description = "A sled based implementation of the `openraft::RaftLogStorage` and `openraft::RaftStateMachine` trait."
+documentation = "https://docs.rs/openraft-sledstore"
 readme = "README.md"
 
 version       = { workspace = true }
@@ -13,12 +13,10 @@ keywords      = { workspace = true }
 license       = { workspace = true }
 repository    = { workspace = true }
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
-openraft = { path= "../openraft", version = "0.9.0", features=["serde"] }
+openraft = { path= "../../openraft", version = "0.10.0", features=["serde"] }
 
-rocksdb = "0.20.1"
+sled = "0.34.7"
 byteorder = "1.4.3"
 serde = { version = "1.0.114", features = ["derive"] }
 serde_json = "1.0.57"
