+++ stores/sledstore/README.md
@@ -1,0 +1,6 @@
+# openraft-sledstore
+
+This is an example `RaftLogStorage` and `RaftStateMachine`  implementation
+with [`sled`](https://github.com/spacejam/sled) based on [openraft-0.9](https://github.com/datafuselabs/openraft/tree/release-0.9).
+
+This crate is built mainly for testing or demonstrating purpose.:)
