+++ stores/sledstore/src/test.rs
@@ -1,6 +1,5 @@
 use std::sync::Arc;
 
-use openraft::storage::Adaptor;
 use openraft::testing::StoreBuilder;
 use openraft::testing::Suite;
 use openraft::StorageError;
@@ -17,8 +16,8 @@ pub fn test_sled_store() -> Result<(), StorageError<ExampleNodeId>> {
     Suite::test_all(SledBuilder {})
 }
 
-type LogStore = Adaptor<TypeConfig, Arc<SledStore>>;
-type StateMachine = Adaptor<TypeConfig, Arc<SledStore>>;
+type LogStore = Arc<SledStore>;
+type StateMachine = Arc<SledStore>;
 
 impl StoreBuilder<TypeConfig, LogStore, StateMachine, TempDir> for SledBuilder {
     async fn build(&self) -> Result<(TempDir, LogStore, StateMachine), StorageError<ExampleNodeId>> {
@@ -26,8 +25,7 @@ impl StoreBuilder<TypeConfig, LogStore, StateMachine, TempDir> for SledBuilder {
 
         let db: sled::Db = sled::open(td.path()).unwrap();
 
-        let store = SledStore::new(Arc::new(db)).await;
-        let (log_store, sm) = Adaptor::new(store);
+        let (log_store, sm) = SledStore::new(Arc::new(db)).await;
 
         Ok((td, log_store, sm))
     }
