+++ tests/tests/metrics/t20_metrics_state_machine_consistency.rs
@@ -56,8 +56,8 @@ async fn metrics_state_machine_consistency() -> Result<()> {
     log_index += 1;
     for node_id in 0..2 {
         router.wait_for_log(&btreeset![node_id], Some(log_index), None, "write one log").await?;
-        let (sto, _sm) = router.get_storage_handle(&node_id)?;
-        assert!(sto.storage().await.get_state_machine().await.client_status.contains_key("foo"));
+        let (_sto, sm) = router.get_storage_handle(&node_id)?;
+        assert!(sm.get_state_machine().await.client_status.contains_key("foo"));
     }
 
     Ok(())
