+++ tests/tests/snapshot_building/t35_building_snapshot_does_not_block_apply.rs
@@ -37,9 +37,7 @@ async fn building_snapshot_does_not_block_apply() -> Result<()> {
     tracing::info!(log_index, "--- set flag to delay snapshot building");
     {
         let (mut _sto1, sm1) = router.get_storage_handle(&1)?;
-        sm1.storage_mut()
-            .await
-            .set_blocking(BlockOperation::DelayBuildingSnapshot, Duration::from_millis(5_000));
+        sm1.block.set_blocking(BlockOperation::DelayBuildingSnapshot, Duration::from_millis(5_000));
     }
 
     tracing::info!(log_index, "--- build snapshot on follower, it should block");
